+++ objects/crypto.go
@@ -53,7 +53,7 @@ type CryptoQuote struct {
 	MarketCap            float64  `json:"marketCap"`
 	PriceAvg50           float64  `json:"priceAvg50"`
 	PriceAvg200          float64  `json:"priceAvg200"`
-	Volume               int64    `json:"volume"`
+	Volume               float64  `json:"volume"`
 	AvgVolume            int64    `json:"avgVolume"`
 	Exchange             string   `json:"exchange"`
 	Open                 float64  `json:"open"`
@@ -61,7 +61,7 @@ type CryptoQuote struct {
 	Eps                  *float64 `json:"eps"`
 	Pe                   *float64 `json:"pe"`
 	EarningsAnnouncement *float64 `json:"earningsAnnouncement"`
-	SharesOutstanding    *float64    `json:"sharesOutstanding"`
+	SharesOutstanding    *float64 `json:"sharesOutstanding"`
 	Timestamp            int64    `json:"timestamp"`
 }
